backend/requirements.txt
altair==4.2.0
anyio==3.6.1
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
configparser==5.3.0
decorator==5.1.1
entrypoints==0.4
etelemetry==0.3.0
fastapi==0.81.0
filelock==3.8.0
PyMuPDF==1.20.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.14.0
looseversion==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
networkx==2.8.6
nibabel==4.0.2
nipype==1.8.3
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pathlib==1.0.1
pdf2image==1.16.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
prov==2.0.0
pyarrow==9.0.0
pydantic==1.9.2
pydeck==0.8.0b1
pydot==1.4.2
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.4
rdflib==6.2.0
requests==2.28.1
rich==12.5.1
scipy==1.9.1
semver==2.13.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.19.1
streamlit==1.12.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
traits==6.4.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
watchdog==2.1.9
wincertstore==0.2
zipp==3.8.1
